+++ src/terndata/ecoplots/_gui/__init__.py
@@ -1,3 +1,5 @@
 from .spatial_selector import spatial_selector
+from .sample_image_viewer import sample_image_viewer
+from .igsn_viewer import igsn_viewer
 
-__all__ = ["spatial_selector"]
+__all__ = ["spatial_selector", "sample_image_viewer", "igsn_viewer"]