+++ docs/changelog.rst
@@ -3,6 +3,14 @@ Change Log
 
 Changes over time for the EcoPlots Python library.
 
+Beta v6
+-------
+
+- Introduced the **has_image** field in sample discovery responses to indicate the presence of associated images.
+- Added the IGSN based sample discovery path to the samples-mode discovery workflow, enabling retrieval of sample metadata based on IGSN values.
+- Added the **samples image viewer** and **samples IGSN viewer**.
+
+
 Beta v5
 -------